OOP + renaming
