+++ geomstats/special_euclidean_group.py
