+++ geomstats/special_orthogonal_group.py
